Move version out further to own VERSION file
