+++ localstack/aws/protocol/service_router.py
@@ -7,7 +7,6 @@ from botocore.model import ServiceModel
 from werkzeug.exceptions import RequestEntityTooLarge
 from werkzeug.http import parse_dict_header
 
-import localstack
 from localstack import config
 from localstack.aws.spec import (
     ServiceCatalog,
@@ -257,7 +256,9 @@ def get_service_catalog() -> ServiceCatalog:
         return ServiceCatalog()
 
     try:
-        ls_ver = localstack.constants.VERSION.replace(".", "_")
+        from localstack.version import VERSION
+
+        ls_ver = VERSION.replace(".", "_")
         botocore_ver = botocore.__version__.replace(".", "_")
         cache_file_name = f"service-catalog-{ls_ver}-{botocore_ver}.pickle"
         cache_file = os.path.join(config.dirs.cache, cache_file_name)
