+++ localstack/constants.py
@@ -1,8 +1,5 @@
 import os
 
-# LocalStack version
-VERSION = "3.2.1.dev"
-
 # HTTP headers used to forward proxy request URLs
 HEADER_LOCALSTACK_EDGE_URL = "x-localstack-edge"
 HEADER_LOCALSTACK_REQUEST_URL = "x-localstack-request-url"
