+++ localstack/version.py
@@ -1,0 +1,1 @@
+VERSION = "3.2.1.dev"
